requirements.txt
absl-py==2.1.0
apex==0.1
attrs==23.2.0
backcall==0.2.0
cachetools==5.4.0

charset-normalizer==3.3.2
cycler==0.11.0
decorator==5.1.1
dominate==2.3.1
einops==0.6.1
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
google-auth==2.32.0
google-auth-oauthlib==0.4.6
grpcio==1.62.2
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1

mkl-service==2.4.0
networkx==2.6.3
ninja==1.11.1.1
numpy==1.21.6
oauthlib==3.2.2
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-ignite==0.1.2
PyYAML==6.0.1
pyzmq==26.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.0.2


tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.9.12
torch==1.13.0+cu116
torch-geometric==2.3.1
torchaudio==0.13.0+cu116
torchfile==0.1.0
torchvision==0.14.0+cu116
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.7
visdom==0.2.4
wcwidth==0.2.13
websocket-client==1.6.1
Werkzeug==2.2.3
yacs==0.1.8
zipp==3.15.0
